Adding repository pattern for application form
Although this is added at this time I haven't fully decided whether
to keep this pattern or not. I've been doing a lot of research on
Ruby on Rails and it doesn't seem like this pattern is commonly used.
Additionally, my goal is for any implementing engineers to be able to
utilize Flex with minimal effort in the future, so more thought and
discussion is needed before this is finalized.

Pros:
- Better separation of concerns
- Easier to test
- Abstracting the data layer and keeping models (that have database
context) from being passed around/overused

Cons:
- More code to write, maintain, and understand