Adding service layer for application form
I added this ApplicationFormService for the purpose of separating the
business logic from the data logic. While the models and repositories
(if the repositories are kept) will be responsible for any data logic,
the service will be responsible for any specific business logic, such
as  the current requirement of checking if a form is already submitted
before allowing any updates to it.

Since this service is where most of the business logic is focused, I
test-drove its creation. I started this in a more "traditional" Ruby on
Rails style (from my understanding), however I also learned about a more
BDD style after this was finished. I may consider switching to that
style in the future.