(Incomplete) Adding application form acceptance test
It's important to note that this commit leaves tests in a partially-
broken state.

My goal here is to add a higher-level acceptance test that will
show ApplicationForm being used how we expect it to be used in the
future. Since ApplicationForm is an abstract class,
BusinessExclusionForm was created to extend from ApplicationForm
and ultimately to be the model under test.

While working on creating the acceptance test, I've created:
- BusinessExclusionForm (the model representing the data)
- BusinessExclusionFormRepository (the repository to manage data interaction)
- BusinessExclusionFormService (the service to manage business logic)

As part of this commit, I've also created a migration for the table
structure, a schema file to define the schema, and a fixture for test
data.

The overall problem that I'm still experiencing is getting fixture data
to properly load on test run. I've restarted this process multiple
times, following Rails instructions (and other guides), but I'm
unfortunately still not able to move forward with this. I plan on
asking for assistance.

Last important note: these files are not how I expect this project to
look in the future. The files are in this state temporarily while I'm
trying a variety of different things to try and get this working.