Adding .gitignore to ignore unversioned files
The .gitignore file from the top level didn't seem to be working for me
so I went ahead and added this one. This seems like a good addition to
make to the Flex engine anyways.

If I should separate this out into its own branch/PR I can do that.