Adding spec directory with RSpec tests
These are the tests converted from the previous `test` directory,
which was using Minitest, to now use RSpec. This also includes
a new factory for fixtures using factory_bot.

Additionally, as part of this commit, I updated some configuration in
`rails_helper.rb` and `spec_helper.rb` to include the necessary updates
to target the spec directory instead of the test directory.