Update passport application form to create case
The PassportApplicationForm, `before_create`, will create a Case
for this application form if it does not already have one. It will then
set the case_id on the application form to the id of the case.

I also updated the foreign key definition in the schema and other
minor cleanups.