Removing business process files and references
This branch was created from the branch
`michael/adding-case-and-business-process-module-pairing-with-loren`.

In this branch, I've removed all files and changes that are unnecessary
in order to add a minimum Case module. If the passport application
case acceptance test still passes without the file or addition, then
I've removed that file or change.