Added event_listener test and logic
I'm TDDing into adding event_listener. Added a test to ensure that an
event listener cannot be added twice. This takes care of a test on
whether the method exists, whether it can be called, and that
the event gets stored. Currently it only allows 1 event to be added,
so I'll have to continue TDDing to get the code to where I want.