Separate application form tests
As the dummy app is getting a little more complex in order to
test case management, I wanted to separate out the tests for
PassportApplicationForm from ApplicationForm.

I made a very simple TestApplicationForm in application_form_spec.rb
to use just in that simple test. This test file only exists to test
ApplicationForm and not any additional functionality that might
be added to the PassportApplicationForm.