Update business process to use callback for cases
When a BusinessProcess needs to find a related Case, it will now use
the given callback instead of a hardcoded type. This allows for more
flexibility and decouples the BusinessProcess from a specific Case type
without doing a slightly odd thing of passing the type in.

This is related to a comment on PR #29 here:
https://github.com/navapbc/template-application-flex/pull/29#discussion_r2044958448