Add view path for task details partials
When trying to render task details partials in the `demo` app, I ran
into a problem with the strictly defined partial route in Flex. Flex had
defined `task/details/{partial name}` as the route to find the
partial...which meant that the parent app would always need to have
their tasks route be in a `task` directory. Adding this view directory
for only the `show` route allows the parent app to name their tasks
directory whatever they want.