Replace composed_of with manual getter/setter definitions (#125)
- Replace composed_of with manual getter/setter definitions
- Subclass ActiveModel::Type::Value instead of ActiveRecord::Type::Value
- Update instructions for contributing Flex attributes

## Context

The Rails composed_of method requires the class the attribute is defined
on to be an ActiveRecord object, which severely limits the applicability
on value objects. This change replaces the composed_of method with
manual definition of getter and setter which has the side benefit of
making the implementation easier to understand with reduced abstractions
and dependencies.

I also realize we don't need to subclass ActiveRecord::Type::Value since
we're not using the serialize/deserialize functionality, so I also
replaced ActiveRecord::Type's with ActiveModel::Type's.