Make staff view navbars the standard size
At present there's some inconsistency with page width, with the staff
navbars having `.maxw-desktop-lg` set, but not on the body. I could see
an argument for the staff pages to have a wider default width (i.e.,
they will generally be using larger displays and interacting with pages
that have more...stuff on them, larger side navs, etc.). But whichever
way we go, should try to keep things consistent. For now, go with the
standard size.