Revert "Return class instead of class name"
This reverts commit f328f56f6fd251671e6a37dd2e7b986adf79e3ef.

The commit was causing the build for downstream projects to fail.
In application.rb the start_listening_for_events call triggers
business process builder which calls start_on_application_created
which eventually gets to the ApplicationForm, which if it has a
has_one_attached call will trigger a call to AWS client new,
which triggers a missing AWS region error.