Remove redundant main tag from tasks index
The tasks index is rednered within the staff layout, which contains a
main tag, making the main within the index unnecessary. This duplicate
tag, which is removed in this commit,  created additional white space
between the navigation bar and the h1 header, leading to visual
inconsistencies. Additionally, having only one main landmark improves
accessibility for users navigating the page with a screen reader.

Partial resolves TS-346